[misc] update obj_id_t to obj_id_type_t; cleaning; solve some int and uint comparison issues

+++ libMimircache/profiler/include/profilerInternal.h
@@ -22,7 +22,7 @@ typedef struct profiler_multithreading_params {
   profiler_res_t *result;
 //  guint64 bin_size;
   GMutex mtx;             // prevent simultaneous write to progress
-  guint64 *progress;
+  gint *progress;
   gpointer other_data;
 } prof_mt_params_t;
